Add support for max_history to Remote
Add two constructor arguments: evt_max_history and tel_max_history.
Add a unit test to check them.
